requirements.txt
openai>=1.0.0
sounddevice>=0.4.6
scipy>=1.11.0
python-dotenv>=1.0.0
numpy>=1.24.0

